move predicates
