+++ package.json
@@ -62,8 +62,10 @@
     "@eslint/js": "^10.0.1",
     "@types/node": "^24.10.1",
     "@vitest/ui": "^4.0.10",
+    "delaunator": "^5.0.1",
     "eslint": "^10.0.1",
     "gh-pages": "^6.3.0",
+    "poly2tri": "^1.5.0",
     "rolldown": "^1.0.0-rc.3",
     "typescript": "^5.0.0",
     "typescript-eslint": "8.56.1-alpha.3",
